Fix build by locking down operating system
Our builds started failing for no obvious reason, turns out a new Debian image
was released in July this year
https://en.wikipedia.org/wiki/Debian_version_history#Debian_10_(Buster)

Lock down the images to Stretch, which is what we were using before.
We can move when ready and not run into seemingly random failures next time an
update gets rolled out. (Although it could be a couple years 😅)