Include database name in the criteria for pooling connections.
Otherwise it is possible for connections to different databases to share underlying MySQL connections. This is problematic as simply referencing the connection via the ActiveRecordHostPool::ConnectionProxy can change the current database out from under another connection.  CONTEXT OF THE ISSUE

This issue presented itself with a combination of using active\_record\_shards and active\_record\_host\_pool. It was representing the sharded and unsharded database connections within the same connection pool. This problem did not manifest itself until Rails 6 when cache clearing for cache dirtying operations—e.g. insert, update, truncate, delete, etc—was updated to loop over the app's connection handlers, connection pools to search for currently active connections.   This happens mid-operation (inside of ActiveRecord::ConnectionAdapters::Mysql2Adapter instance) just before the operation is carried out. The result was that referencing the connection was changing the currently selected database and it was not being restored. This led to issues where the operation would try to be carried out on a different database that did not have the same tables/columns.
 REFERENCES

The problematic line that causes the database to switch is: https://github.com/zendesk/active\_record\_host\_pool/blob/981fdeb8dfa52f390bd91471adc19049283c6fb8/lib/active\_record\_host\_pool/connection\_proxy.rb#L16

The change in Rails 6 that exercised a code path to expose this surface the issue: https://github.com/rails/rails/pull/35089/files#diff-193e6e4a86bf5d360a5001bce59334cbR180-R186

WHY THIS FIX

Connection (aka resource) pools are generally used to recycle connections that can be used within the same context. For the purposes of what ActiveRecordHostPool does context seems to mean databases with the same tables/columns/etc.

However, connections that point to different databases on the same host are unlikely to represent similar contexts. This change-set moves to consider them entirely different contexts. By doing so this circumvents any issues where higher-level connections that point to different databases will share an underlying mysql connection.

BENEFITS OF THE PREVIOUS IMPLEMENTATION

Prior to this change, the implementation reduced the number of connections by switching the database using SELECT DATABASE SQL statements.

This has worked for nearly a decade, but has always been open to the issue of one connection changing the database out from under an operation assuming a different database. It just so happens that Rails 6 begins to exercise a code path that exposes this issue.

BENEFITS OF THE NEW IMPLEMENTATION

By pooling connections together with criteria that ensures interchangeable contexts we eliminate the chance for this issue to exist.

CONSEQUENCES OF THE NEW IMPLEMENTATION

This will increase the number of connection pools. Previously it was based on the number of unique database combinations of host+port+socket+username+slave. Now, it will also include database name.

This will also likely increase the number of open database connections as these pools will now each have their own number of connections limit to manage.