Patch clear_on_handler
When Rails calls `#clear_query_caches_for_current_thread` we end up in
`#clear_on_handler`.

Clear on handler will loop through every connection pool (but really
every pool proxy because ARHP creates proxiws) and call
`pool.connection` on it. `pool.connection` will cause ARHP to call
`select_db` on that selection and thus we're switching the current
database on all connections when in the process of clearing the cache.

We can then end up currently on a database we don't expect and therefore
we switch back to avoid conflicts.

This patch broke a caching test because it's solving the problem before
we testing that the patches work. I've made that test's setup more
robust to recreate the original issue, confirm it works with the patch,
then confirm it works without the patch.