Fix ThreadSafetyTest
When including a module into a class defined with `class Foo`, the constants
defined in the module can be accessed with "implicit" namespace. When the class
is defined with `Class.new`, the same constants can only be access with explicit
namespace. An example:

    module M
      X = 1
    end

    class A
      include M
      def foo
        X
      end
    end

    B = Class.new do
      include M
      def foo
        X
      end
    end

    >> A.new.foo
    => 1

    >> B.new.foo
    Traceback (most recent call last):
      ...
    (irb):15:in `foo': uninitialized constant X (NameError)

The solution chosen is to use `class Foo` syntax, wrap the whole thing in an
`unless` statement and just try to live with the extra indentation.