Drop `reconnect: true` from DB configs
We get flooded with these errors in tests:

```
WARNING: MYSQL_OPT_RECONNECT is deprecated and will be removed in a future version.
```

`reconnect: true` allows for automatic reconnection to the server in
the event that the server disconnects the client. One way that happens
is if the connection sits unused long enough to exceed the session
timeout. ARHP doesn't manage reconnection on its own so we can remove this.

This did require manually adding a `connection.reconnect!` to a test
where we tell the server to kill the connection. In that test we're
testing that calling `reconnect!` doesn't cause us to switch to the
wrong database, not the specific behavior of `reconnect: true`.