Add the lockfiles back
We've decided that lockfiles are useful to include because they give us
a combination of gem versions that we know work. If we try updating the
lockfiles and the tests start failing we at least have baseline we can
fall back to and then update gems one-by-one to figure out which gem
breaks things.

Plus, we intend to get this gem testing with Rails main. That won't have
a lockfile checked in so it will serve as our canary.