Add support for Rails 7.2 & main
Necessary changes:

\#\# Rename `@connection` -> `@raw_connection` adapter mixin.

`@connection` was renamed to `@raw_connection` on the database adapters
requiring us to conditionally define the `#_real_connection_object_id`
method. One returns `@connection.object_id` and the other
`@raw_connection.object_id`.

\#\# Rename `#connection` -> `#lease_connection` on the PoolProxy.

`#connection` was renamed to `#lease_connection` in Rails[^1]

\#\# Refactor `#with_connection` on the PoolProxy.

`#with_connection` in Rails now leverages a new `ConnectionPool::Lease`
class[^2]. Essentially, the *real* connection pool has a connection
lease for the thread that may, or may not, have a *real* connection that may or
may not be "sticky". See the details about what that means in the Rails
commit. For ARHP purposes, we are essentially implementing the same
logic of `#with_connection` in Rails. We ask the real pool for its
`connection_lease`. We ask the lease for its *real* connection _or_
checkout a new *real* connection. Once we have a *real* connection we
yield a connection *proxy* for that *real* connection, checking it back
in unless the connection lease is "sticky".

\#\# Memoize a PoolProxy in the `schema_cache`.

ConnectionPool passes itself to a `BoundSchemaReflection` which it
stores in its `@schema_cache` ivar[^3]. The pool is set as `@pool` on
the `BoundSchemaReflection` which will end up calling `with_connection`
on the pool[^4]. If this is a _real_ pool and not a PoolProxy we can end
up using the wrong database.

[^1]:(https://github.com/rails/rails/commit/38e8609ded6cf01f6ed20b1134499e07cd68293f#diff-642b90553b888bd2c724c093a1a685a5408a7d8293f3751366c25dc548936eb7R290)
[^2]:(https://github.com/rails/rails/commit/38e8609ded6cf01f6ed20b1134499e07cd68293f#diff-642b90553b888bd2c724c093a1a685a5408a7d8293f3751366c25dc548936eb7R371)
[^3]:(https://github.com/rails/rails/blob/fa048105d145536538b923fa57465bdeee559e88/activerecord/lib/active_record/connection_adapters/abstract/connection_pool.rb#L266)
[^4]:(https://github.com/rails/rails/blob/fa048105d145536538b923fa57465bdeee559e88/activerecord/lib/active_record/connection_adapters/schema_cache.rb#L320)