Move _switch_connection into with_raw_connection
Right now `ActiveRecordHostPool` never makes it into ActiveRecord's
reconnect/retry logic. That's because that logic starts inside of the
`#with_raw_connection` method[^1]. Here's a simplified look at the method:

```ruby
def with_raw_connection(allow_retry: false, materialize_transactions: true)
  # ...
  begin
    yield @raw_connection # <-- The raw connection is yielded to the block.
  rescue => exception
    # ...
    if reconnectable && retryable_connection_error?(exception)
      # Reconnect & retry certain errors.
      reconnect!
      retry
    end
  end
end
```

In host pool we patch `#raw_execute` to conduct the database switch:

```ruby
def raw_execute(...)
  if _host_pool_desired_database && !_no_switch
    _switch_connection
  end
  super
end
```

`#raw_execute` in Rails wraps the query in `#with_raw_connection` and
thus that's where we get the reconnect behavior:

```ruby
def raw_execute(sql, name, async: false, allow_retry: false, materialize_transactions: true)
  # ...
  with_raw_connection(allow_retry: allow_retry, materialize_transactions: materialize_transactions) do |conn|
    result = conn.query(sql)
  end
  # ...
end
```

Because we're attempting to switch databases outside of the
`#with_raw_connection`, if that database switch fails then we won't reconnect or retry.

A better solution would be to instead patch `#with_raw_connection`.
Because that method yields the `raw_connection` (which we use to call
`select_db` on) we can patch it so that we call `select_db` on the
yielded raw connection first and then execute the rest of the block.

An added bonus to this is that we're no longer calling `#raw_connection`
anywhere. `#raw_connection` would mark the connection "dirty"[^2]
forcing us to `#clean!` and verify it every time. We're using the
`raw_connection` that's yielded by `#with_raw_connection` keeps the
connection "clean".

Three other notes:

When we instantiate a `ConnectionProxy` we used to call `raw_execute("SELECT
1")` to trigger a database switch. I removed that because `raw_execute`
is where we pass `allow_retry` and I don't think ActiveRecordHostPool
should make a decision about if we're allowing retries or not. So
instead of instantiating a `ConnectionProxy` and immediately forcing a
`select_db` we wait to switch databases until the first time
`with_raw_connection` is called.

Related to the above: `test_arhp_wrong_db.rb` is updated to call
`TestNotThere.connection.execute("SELECT 1")` instead of just
`TestNotThere.connection`. Now that we're lazily switching databases we
don't trigger a `SELECT 1` call just by instantiating a new connection.
This means we have to explicitly trigger a database query in this test.

In `test_thread_safety.rb` we would conditionally call `#connection.raw_execute(...).first.first` or
`#connection.execute(...).first.first` depending on what version of
Rails and which adapter we're using. That's because, depending on the
implementation, `execute` would either return the value returned from
the query _or_ an array with the column name and the returned value.
Because we only care about the value in these tests I've changed it so
that we call `connection.query_value(...)` which does exactly what we want.

[^1]:(https://github.com/rails/rails/blob/a86a88965bfa4d596156b08901164be2f0d0da6c/activerecord/lib/active_record/connection_adapters/abstract_adapter.rb#L975-L1004)

[^2]:(https://github.com/rails/rails/blob/a86a88965bfa4d596156b08901164be2f0d0da6c/activerecord/lib/active_record/connection_adapters/abstract_adapter.rb#L790-L796)