Remove dependency on mutex_m
Mutex_m used to be in stdlib, but in newer Ruby versions it is
extracted to a separate gem.

We only used the synchronize method, and it's just as easy to call it
directly on a Thread::Mutex instance. As you can see in
https://github.com/ruby/mutex_m/blob/v0.2.0/lib/mutex_m.rb it saves us
from exposing a few other methods onto our object.