Allow working with Mysql2 and Trilogy at same time
Previously, ARHP would check if `mysql2` was in the loaded specs and
then patch `mysql2`. If it wasn't part of the loaded specs, it would try Trilogy.

This limited ARHP to work with one, or the other, but not both. We can
actually bundle `mysql2` and `trilogy` in the same gemfile and use them
side-by-side, conditionally enabled them, etc. This changes it so that
ARHP tries loading and patching both adapters.