Delete all records instead of dropping dbs
We have a race condition that makes some of our tests flaky.

Previously we would call `Phenix.burn!` after _each_ test to drop all of
the databases and then recreate them after. This cleared all of the
records but also helped us with some tests where we expected some
databases to not exist.

The problem with this is that ActiveRecord's connection pool reaper
will, by default, start to kill connections after they're idle for 60
seconds and then reconnect[^1]. It also does this in a thread [^2].

What used to occasionally happen is that the reaper would kill a
connection and then attempt to reconnect to the database _in between
test runs_. i.e. _after_ the databases were dropped but _before_ they
were created again. This would blow up.

This commit changes it so we no longer drop the databases but instead
call `delete_all` on every model.

This required some other changes:

1. Call `ActiveRecord::Base.connection.execute(...)` in `test_arhp.rb`
to ensure we've established a connection before we assert that it's
connected. (Phenix.rise! would establish a connection previously and we
intentionally call `disconnect!` after each test).
2. Update the `test_arhp_wrong_db.rb` tests to use configurations with
the right credentials but made up database names so that we raise
"database not found" errors. We used to drop all of the databases before
this test but now we don't.

[^1]:(https://github.com/rails/rails/blob/a86a88965bfa4d596156b08901164be2f0d0da6c/activerecord/lib/active_record/connection_adapters/abstract/connection_pool/reaper.rb#L15-L16)

[^2]:(https://github.com/rails/rails/blob/a86a88965bfa4d596156b08901164be2f0d0da6c/activerecord/lib/active_record/connection_adapters/abstract/connection_pool/reaper.rb#L42)