Stop overriding #class
ARHP has been overriding `#class` so that `connection.class.class_eval`
can work. That makes it _very_ hard to debug ARHP issues because you're
never quite sure what object you're working with.

I've changed it so that we now raise an exception if you try to call
`.class_eval` on the `ConnectionProxy` class but implements a
`._class_eval` method that will allow the typical `class_eval` behavior
if someone _really_ needs to.